# docs/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Blink Backend
  version: 0.0.0
  description: API para encurtar URLs.
...
paths:
  /users:
    post:
      summary: Criar usuário
      operationId: users/create
      tags:
        - Usuários
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                name:
                  type: string
                  description: O nome do usuário
                  examples:
                    - Diego Aquino
                email:
                  type: string
                  format: email
                  description: O email do usuário
                  examples:
                    - diego@email.com
                password:
                  type: string
                  description: A senha do usuário
                  examples:
                    - 123456
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '409':
          description: Email em uso por outro usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: O id do usuário
        schema:
          type: string
        example: tz4a98xxat96iws9zmbrgj3a

    get:
      summary: Buscar usuário
      operationId: users/getById
      tags:
        - Usuários
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '403':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

    patch:
      summary: Atualizar usuário
      operationId: users/update
      tags:
        - Usuários
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: O nome do usuário
                  examples:
                    - Diego Aquino
                email:
                  type: string
                  format: email
                  description: O email do usuário
                  examples:
                    - diego@email.com
      responses:
        '200':
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '403':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '409':
          description: Email em uso por outro usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
  /{redirectId}:
    parameters:
      - name: redirectId
        in: path
        required: true
        description: O id do redirecionamento
        schema:
          type: string
        example: t96iws9z

    get:
      summary: Redirecionar
      operationId: redirect
      tags:
        - Redirecionamentos
      responses:
        '308':
          description: Redirecionamento
          headers:
            location:
              schema:
                type: string
                examples:
                  - https://google.com
            cache-control:
              schema:
                type: string
                examples:
                  - public, max-age=0, must-revalidate
        '404':
          description: Redirecionamento não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
components:
  schemas:
    User:
      type: object
      required:
        - id
        - name
        - email
        - type
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: O id do usuário
          examples:
            - tz4a98xxat96iws9zmbrgj3a
        name:
          type: string
          description: O nome do usuário
          examples:
            - Diego Aquino
        email:
          type: string
          format: email
          description: O email do usuário
          examples:
            - diego@email.com
        type:
          type: string
          description: O tipo do usuário
          enum:
            - SYSTEM_ADMINISTRATOR
            - NORMAL
          examples:
            - NORMAL
        createdAt:
          type: string
          format: date-time
          description: A data de criação do usuário
          examples:
            - '2021-08-01T00:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          description: A data de atualização do usuário
          examples:
            - '2021-08-01T00:00:00.000Z'
...
    ValidationError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: A mensagem de erro
          examples:
            - Validation error
        issues:
          type: array
          description: Os problemas de validação
          items:
            type: object
            properties:
              message:
                type: string
                description: A mensagem de erro
                examples:
                  - 'Invalid input: expected string, received number'
              code:
                type: string
                description: O código do erro
                examples:
                  - invalid_type
              path:
                type: array
                description: O caminho do erro
                items:
                  oneOf:
                    - type: string
                    - type: number
                examples:
                  - - name
            additionalProperties: true
    AuthError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: A mensagem de erro
          examples:
            - Unauthorized
    NotFoundError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: A mensagem de erro
          examples:
            - Not found
    ConflictError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: A mensagem de erro
          examples:
            - Conflict
    InternalServerError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: A mensagem de erro
          examples:
            - Internal server error
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT